requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
pycryptodome>=3.19.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
